+++ src/ModelUuidServiceProvider.php
@@ -3,6 +3,8 @@
 namespace Touhidurabir\ModelUuid;
 
 use Illuminate\Support\ServiceProvider;
+use Touhidurabir\ModelUuid\UuidGenerator\Generator;
+use Touhidurabir\ModelUuid\Console\RegenerateModelUuid;
 
 class ModelUuidServiceProvider extends ServiceProvider {
     
@@ -16,6 +18,11 @@ class ModelUuidServiceProvider extends ServiceProvider {
         $this->mergeConfigFrom(
             __DIR__.'/../config/model-uuid.php', 'model-uuid'
         );
+
+        $this->app->bind('model-uuid', function () {
+            
+            return new Generator;
+        });
     }
 
 
@@ -25,6 +32,12 @@ class ModelUuidServiceProvider extends ServiceProvider {
      * @return void
      */
     public function boot() {
+
+        if ( $this->app->runningInConsole() ) {
+			$this->commands([
+				RegenerateModelUuid::class
+			]);
+		}
 
         $this->publishes([
             __DIR__.'/../config/model-uuid.php' => base_path('config/model-uuid.php'),