+++ CHANGELOG.md
@@ -21,3 +21,10 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.4] - 2021-09-22
 - Added default on missing config values.
+
+## [1.1.0] - 2021-10-29
+- Bug fix.
+- Regeneration command and job added.
+- New Features like temporarily disable/enable added.
+- Tests update.
+- README update.